+++ src/auth/session/storage/custom.ts
@@ -10,10 +10,16 @@ export class CustomSessionStorage implements SessionStorage {
       id: string,
     ) => Promise<SessionInterface | {[key: string]: unknown} | undefined>,
     readonly deleteCallback: (id: string) => Promise<boolean>,
+    readonly deleteSessionsCallback?: (ids: string[]) => Promise<boolean>,
+    readonly findSessionsByShopCallback?: (
+      shop: string,
+    ) => Promise<SessionInterface[]>,
   ) {
     this.storeCallback = storeCallback;
     this.loadCallback = loadCallback;
     this.deleteCallback = deleteCallback;
+    this.deleteSessionsCallback = deleteSessionsCallback;
+    this.findSessionsByShopCallback = findSessionsByShopCallback;
   }
 
   public async storeSession(session: SessionInterface): Promise<boolean> {
@@ -76,5 +82,41 @@ export class CustomSessionStorage implements SessionStorage {
         `CustomSessionStorage failed to delete a session. Error Details: ${error}`,
       );
     }
+  }
+
+  public async deleteSessions(ids: string[]): Promise<boolean> {
+    if (this.deleteSessionsCallback) {
+      try {
+        return await this.deleteSessionsCallback(ids);
+      } catch (error) {
+        throw new ShopifyErrors.SessionStorageError(
+          `CustomSessionStorage failed to delete array of sessions. Error Details: ${error}`,
+        );
+      }
+    } else {
+      console.warn(
+        `CustomSessionStorage failed to delete array of sessions. Error Details: deleteSessionsCallback not defined.`,
+      );
+    }
+    return false;
+  }
+
+  public async findSessionsByShop(shop: string): Promise<SessionInterface[]> {
+    let sessions: SessionInterface[] = [];
+
+    if (this.findSessionsByShopCallback) {
+      try {
+        sessions = await this.findSessionsByShopCallback(shop);
+      } catch (error) {
+        throw new ShopifyErrors.SessionStorageError(
+          `CustomSessionStorage failed to find sessions by shop. Error Details: ${error}`,
+        );
+      }
+    } else {
+      console.warn(
+        `CustomSessionStorage failed to find sessions by shop. Error Details: findSessionsByShopCallback not defined.`,
+      );
+    }
+    return sessions;
   }
 }